requirements.txt
requests
pandas
matplotlib
Flask
PyYAML
gunicorn
